requirements.txt
quart==0.20.0
requests==2.32.4
aiohttp==3.10.11
hypercorn==0.16.0
imageio==2.34.0
imageio-ffmpeg==0.4.9
Pillow==10.3.0